Add missing include for GenParticle to GenVertex.h
Right now GenVertex.h doesn't compile when included because we
instantiate the GenParticlePtr template in this header which
requires the definition of GenParticle to be available. However,
we only have a forward declaration available at this point, so
we get compilation errors about a missing definition.

This patch just adds a include to GenParticle.h which fixes this.