More LH17 changes plus HeavyIon stuff
Together with Jan Fiete I have updated and improved documentation of the
GenHeavyIon class. Several things have been made deprecated and a
versioning for the i/o was introduced.

The usage of multi-file runs in LHEF now seems to work, and the
LHEFExample directory has been modified to use this new facility.