Fix the memcheck tests.
The variables are defiend as MEMORYCHECK_COMMAND{,_OPTIONS} and then
used as CMAKE_MEMORYCHECK_COMMAND{,_OPTIONS} which means they are
ignored, and the memcheck simply repeats the original test unmodified.

The default value for MEMORYCHECK_COMMAND_OPTIONS is invalid.
--leak-check=full is an option to the memcheck tool and can not be
used together with the --tool=callgrind option. The mamcheck tests
fail with "valgrind: Unknown option: --leak-check=full"