The memcheck and non-memcheck version of the same test write to the same files and can not be run in parallel, or the tests will fail. This adds a dependency so that the memcheck version can not start before the non-memcheck version has finished.
The setting of the TIMEOUT option here sets the timeout for the
non-memcheck version of the test, which is redundand since it is
already set before. It could be possible to set the timeout for the
memcheck version here, i.e. changing it to

set_tests_properties(memcheck_${test} PROPERTIES TIMEOUT xx)

instead of removing it, but the timeout value must be increased then.
(I tried 60 instead of 10, but some of the tests still failed due to
timeout with this value, and it takes even longer on slpwer
architectures).